Add logic to use slightly less IMAD, SHF, and LEA instructions
Add logic to use slightly less IMAD, SHF, and LEA instructions. The
IADD3 instruction increased though. But overall this looks like an
improvement in the number of instructions executed. But we need to do
this for the shared memory buffer as well as the `memcpy_async`
instruction will only be issued when both the SMEM and Global addresses
are available.